Fixes to Contribution Guidelines section of README.md
Fenced code blocks for items #6 and #7 were not rendering properly, fixed. All code blocks in this section used "javascript" as the language for syntax highlighting, fixed to "shell" since the code consists of CLI commands, not javascript. Changed order of ### and list numbers (example: from 1. ### to ### 1) so that the numbers are styled properly.